pyproject.toml
[project]
name = "proxyhostslist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keyring>=25.6.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.45.0",
]
